Using firebase emulators for storage & firestore
The firebase emulators provide a more integrated
dev experience, with an emulator UI that makes
it easy to inspect the firestore contents in particular.

However, the firebase storage emulator is only
partially compatible with the cloud storage API.
It needs a couple of very particular workarounds
to make the same logic work both against cloud APIs
and the emulator.