Push error identification back to API call sites
The Firestore APIs allow us to classify get operations immediately when called, but write operations will only
return errors when the transaction is committed.
This made the database helpers 'internal API' with error
types inonsistent.

Also, the error returned from a failed commit does not
contain enough details (at least not without significant
unpacking) to establish which of the writes failed.

Because of this, it's better to expose the gRPC style error API
to the call sites, i.e. detecting certain types of
errors with `if status.Code(err) == codes.NotFound { ... }` style
checks.