Maintain store-files & store-file-hashes in DB
In addition to keeping a log of the uploads,
the DB also tracks which files and corresponding
hashes exist. This helps for browsing, and
can also help for expiring blobs when all
referencing uploads expire.

This is not working entirely yet; the store-files and
store-file-hashes should only represent blobs
that are currently active. Uploads should present
the entire history. Right now, store-files
and store-file-hashes will live forever.