Abort will abort already-uploaded files
The 'abort' operation will mark all files in the upload
as aborted, even if some of them completed upload.
This makes the upload operation closer to atomic
from a DB point of view.

Aborts will result in some files in the file storage
that are not tracked in the DB. That should not
be a practical problem, except that people
can consume lots of storage space by doing lots
of aborted uploads of unique files.