Upgraded to Node 16.19.0 + latest deps
We'd like to go to Node v17, but there's a
breaking change that forces up to update
babel-loader to 9.0.0 or newer; unfortunately
cli-plugin-typescript is locked to babel-loader
8.2.2 and there's no way around tha except
forking / internalizing cli-plugin-typescript
and manually changing that dependency.
Not worth it, so we stay on Node v16
for the time being.

Firebase usage uses the v9 compat
libraries and should be changed to v9
modular libraries.