Renamed hash => blob
"Hash" sounded like it only referred to
the hash of a file, and not the file itself.
Blob is a better word for an arbitrary file.

This breaks backward compatibility with
0.4 clients that want to upload things.
This can be fixed by making the upload
API calls backward compatible (accepting
and returning both Hash and
DebugIdentifier parameters).